Eventy

+++ src/main/java/com/dominik/typer/model/User.java
@@ -5,8 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @AllArgsConstructor
@@ -15,6 +13,6 @@ public class User {
     private String username;
     private Integer points;
     private String email;
-    private BigDecimal balance;
+    private Double balance;
     private UserRole userType ;
 }
